Revamp module system to avoid getters and heap allocation

+++ libsimulator/include/world/world_forward.h
@@ -9,7 +9,7 @@ struct dxBody;
 //typedef struct dxWorld *world_id; /* dWorldID; */
 typedef struct dxBody *world_body;   /* dBodyID; */
 
-typedef uint64_t coord;
+typedef int coord;
 
 
 #endif
